Remove (broken) Docker-based Python LSP
This never actually worked. Plus, as noted in #48, the Homebrew recipe for
`python-lsp-server` installs several (not all) of my desired extensions
alongside the main server itself. Assuming the Homebrew "version" of
`python-lsp-server` "works" (e.g. does jump-to-def'n for virtualenv packages as
well), then that obviates the need for this bespoke Docker-language-server thing
entirely.